pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "k1nderko",email = "107281239+k1nderko@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "db-py (>=0.5.3,<0.6.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
